Refactor Command class and update usage in Program.cs
Significantly redefined the `Command` class in `Command.cs`, adding detailed XML documentation for methods `CreateRelease`, `DownloadAsset`, and `UploadAsset`. Updated `using` directives to include necessary namespaces while removing obsolete ones.

In `Program.cs`, simplified method calls to `Command.CreateRelease` and `Command.DownloadAsset` by removing the namespace prefix, aligning with the new structure of the `Command` class.
